Refactor: move storage related types into types/v070. Dir "types" includes all data-compatibility related types
