+++ openraft/src/store_ext.rs
@@ -5,17 +5,17 @@ use std::sync::RwLock;
 
 use crate::async_trait::async_trait;
 use crate::raft::Entry;
-use crate::storage::HardState;
-use crate::storage::LogState;
-use crate::storage::Snapshot;
 use crate::summary::MessageSummary;
+use crate::types::v070::log_state::LogState;
 use crate::AppData;
 use crate::AppDataResponse;
 use crate::DefensiveCheck;
 use crate::EffectiveMembership;
+use crate::HardState;
 use crate::LogId;
 use crate::RaftStorage;
 use crate::RaftStorageDebug;
+use crate::Snapshot;
 use crate::SnapshotMeta;
 use crate::StateMachineChanges;
 use crate::StorageError;
