+++ openraft/src/testing/suite.rs
@@ -6,16 +6,16 @@ use maplit::btreeset;
 
 use crate::raft::Entry;
 use crate::raft::EntryPayload;
-use crate::storage::HardState;
-use crate::storage::InitialState;
-use crate::storage::LogState;
 use crate::testing::DefensiveStoreBuilder;
 use crate::testing::StoreBuilder;
+use crate::types::v070::log_state::LogState;
 use crate::AppData;
 use crate::AppDataResponse;
 use crate::DefensiveError;
 use crate::EffectiveMembership;
 use crate::ErrorSubject;
+use crate::HardState;
+use crate::InitialState;
 use crate::LogId;
 use crate::Membership;
 use crate::RaftStorage;
