+++ openraft/src/types/v070/effective_membership.rs
@@ -1,0 +1,20 @@
+use serde::Deserialize;
+use serde::Serialize;
+
+use super::LogId;
+use super::Membership;
+
+/// The currently active membership config.
+///
+/// It includes:
+/// - the id of the log that sets this membership config,
+/// - and the config.
+///
+/// An active config is just the last seen config in raft spec.
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct EffectiveMembership {
+    /// The id of the log that applies this membership config
+    pub log_id: LogId,
+
+    pub membership: Membership,
+}
