+++ openraft/src/types/v070/hard_state.rs
@@ -1,0 +1,16 @@
+use serde::Deserialize;
+use serde::Serialize;
+
+use super::NodeId;
+
+/// A record holding the hard state of a Raft node.
+///
+/// This model derives serde's traits for easily (de)serializing this
+/// model for storage & retrieval.
+#[derive(Clone, Debug, Serialize, Deserialize, Eq, PartialEq, Default)]
+pub struct HardState {
+    /// The last recorded term observed by this system.
+    pub current_term: u64,
+    /// The ID of the node voted for in the `current_term`.
+    pub voted_for: Option<NodeId>,
+}
