+++ openraft/src/types/v070/snapshot.rs
@@ -1,0 +1,38 @@
+use serde::Deserialize;
+use serde::Serialize;
+use tokio::io::AsyncRead;
+use tokio::io::AsyncSeek;
+
+use super::LogId;
+
+// Everytime a snapshot is created, it is assigned with a globally unique id.
+pub type SnapshotId = String;
+
+/// The identity of a segment of a snapshot.
+#[derive(Debug, Default, Clone, PartialOrd, PartialEq, Eq, Serialize, Deserialize)]
+pub struct SnapshotSegmentId {
+    pub id: SnapshotId,
+    pub offset: u64,
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
+pub struct SnapshotMeta {
+    // Log entries upto which this snapshot includes, inclusive.
+    pub last_log_id: LogId,
+
+    /// To identify a snapshot when transferring.
+    /// Caveat: even when two snapshot is built with the same `last_log_id`, they still could be different in bytes.
+    pub snapshot_id: SnapshotId,
+}
+
+/// The data associated with the current snapshot.
+#[derive(Debug)]
+pub struct Snapshot<S>
+where S: AsyncRead + AsyncSeek + Send + Unpin + 'static
+{
+    /// metadata of a snapshot
+    pub meta: SnapshotMeta,
+
+    /// A read handle to the associated snapshot.
+    pub snapshot: Box<S>,
+}
